still o(n-i) but better:

```
***** model forward in  57.86 ms
***** model schedule in  18.09 ms
***** model lower in 2498.97 ms
```

```
***** model forward in  57.45 ms
***** model schedule in  17.80 ms
***** model lower in 1797.98 ms
```